course spreadsheet data, move stuff around
