+++ resources/js/features/course-planning/helpers/getCourseTableRows.ts
