+++ resources/js/features/course-planning/helpers/getEnrollmentsForCourseInTerm.ts
@@ -1,5 +1,6 @@
 import * as T from "@/types";
 import { getJoinedEnrollmentRecord } from "./getJoinedEnrollmentRecord";
+import { isNotNullish } from "@/utils/isNotNullish";
 
 export function getEnrollmentsForCourseInTerm({
   course,
@@ -11,12 +12,18 @@ export function getEnrollmentsForCourseInTerm({
   lookups: T.CoursePlanningLookups;
 }): T.JoinedEnrollmentRecord[] {
   return Object.values(lookups.enrollmentLookup)
-    .map((enrollment) =>
-      getJoinedEnrollmentRecord({
-        enrollment,
-        lookups,
-      }),
-    )
+    .map((enrollment) => {
+      // if a person is not in the person lookup, then we can't join the enrollment and so we skip it
+      try {
+        return getJoinedEnrollmentRecord({
+          enrollment,
+          lookups,
+        });
+      } catch (e) {
+        return null;
+      }
+    })
+    .filter(isNotNullish)
     .filter((joinedEnrollment) => {
       const section =
         lookups.sectionLookup[joinedEnrollment.enrollment.sectionId];
