+++ resources/js/features/course-planning/helpers/getTermsWithLeaves.ts
@@ -1,16 +1,15 @@
 import * as T from "@/types";
 import { getLeavesInTerm } from "./getLeavesInTerm";
 
-export function getLeavesRow({
+export function getTermsWithLeaves({
   lookups,
   filters,
 }: {
   lookups: T.CoursePlanningLookups;
   filters?: T.CoursePlanningFilters;
-}): T.LeaveRow {
-  const termLeaves = Object.values(lookups.termLookup).map((term) => {
+}): T.TermLeaves[] {
+  return Object.values(lookups.termLookup).map((term) => {
     const leaves = getLeavesInTerm({ lookups, term });
     return { term, leaves };
   });
-  return ["leaves", ...termLeaves];
 }
