+++ resources/js/features/course-planning/stores/index.ts
@@ -4,4 +4,4 @@ export { useEnrollmentStore } from "./useEnrollmentStore";
 export { useCourseStore } from "./useCourseStore";
 export { useCourseSectionStore } from "./useCourseSectionStore";
 export { useTermStore } from "@/stores/useTermStore";
-export { useRootCoursePlanningStore } from "./useRootCoursePlanningStore";
+export { useCoursePlanningStore } from "./useCoursePlanningStore";
