+++ resources/js/utils/isNotNullish.ts
@@ -1,0 +1,3 @@
+export function isNotNullish<T>(value: T | null | undefined): value is T {
+  return value !== null && value !== undefined;
+}
